teste 5

+++ ecommerce/src/main/java/br/edu/unifio/ecommerce/repositorios/PagamentoRepositorio.java
@@ -3,6 +3,6 @@ package br.edu.unifio.ecommerce.repositorios;
 import org.springframework.data.jpa.repository.JpaRepository;
 import br.edu.unifio.ecommerce.entidades.Pagamento;
 
-public interface PagamentoRepositorio extends JpaRepository<Pagamento, Short> {
+public interface PagamentoRepositorio extends JpaRepository<Pagamento, Integer> {
 
 }
